Initial Commit. First working Version
Commit first alpha version with working system. System needs to be refactored on many places because it damages clean code principles to prevent potential damage in future. Not working: naming of generated files, preview of pictures, writing in file to check if image is unique, settings on blueprint (in preview), etc..